// src/app/core/api/generated/model/procedure.ts
/**
 * VetClinicAPI_V2 | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetAnimalByIdIdParameter } from './getAnimalByIdIdParameter';


export interface Procedure { 
    id?: GetAnimalByIdIdParameter;
    code: string;
    name: string;
    isDeleted?: boolean;
}